refactor(server): extract routes into routers module and clean up imports
- Add web/server/routers/ with prediction.py and results.py, moving
  route handlers out of main.py for better separation of concerns
- Refactor main.py to register APIRouter instances instead of inline
  route definitions (net -68 lines)
- Reformat imports across prediction_service, batch_price_prediction,
  ohlcv_data, and save_results to isort style
- Minor formatting fixes in saved_result.py schema (blank line spacing)
- Fix import order in test_prediction_api_errors.py
- Minor fix in frontend handleAPI.ts
